+++ src/game/Object/PlayerLoad.cpp
@@ -1,0 +1,2331 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+/**
+ * @file PlayerLoad.cpp
+ * @brief Cohesion split of Player.cpp -- the LOAD SYSTEM block:
+ *        `_Load*` DB helpers, `LoadFromDB`, instance-binding lifecycle,
+ *        and the raid-info/saved-instance packet senders. Same `Player`
+ *        class; no behaviour change. CMake `file(GLOB Object/*.cpp)`
+ *        picks this file up automatically; Player.h is unchanged.
+ */
+
+#include "Player.h"
+#include "Language.h"
+#include "Database/DatabaseEnv.h"
+#include "Log.h"
+#include "Opcodes.h"
+#include "SpellMgr.h"
+#include "World.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "UpdateMask.h"
+#include "SkillDiscovery.h"
+#include "QuestDef.h"
+#include "GossipDef.h"
+#include "UpdateData.h"
+#include "Channel.h"
+#include "ChannelMgr.h"
+#include "MapManager.h"
+#include "MapPersistentStateMgr.h"
+#include "InstanceData.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "CellImpl.h"
+#include "ObjectMgr.h"
+#include "ObjectAccessor.h"
+#include "CreatureAI.h"
+#include "Formulas.h"
+#include "Group.h"
+#include "Guild.h"
+#include "GuildMgr.h"
+#include "Pet.h"
+#include "Util.h"
+#include "Transports.h"
+#include "Weather.h"
+#include "BattleGround/BattleGround.h"
+#include "BattleGround/BattleGroundMgr.h"
+#include "BattleGround/BattleGroundAV.h"
+#include "OutdoorPvP/OutdoorPvP.h"
+#include "ArenaTeam.h"
+#include "Chat.h"
+#include "revision_data.h"
+#include "Database/DatabaseImpl.h"
+#include "Spell.h"
+#include "ScriptMgr.h"
+#include "SocialMgr.h"
+#include "AchievementMgr.h"
+#include "Mail.h"
+#include "SpellAuras.h"
+#include "DBCStores.h"
+#include "DB2Stores.h"
+#include "SQLStorages.h"
+#include "Vehicle.h"
+#include "Calendar.h"
+#include "DisableMgr.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+// corpse reclaim times -- file-local in Player.cpp; mirrored here so
+// LoadFromDB can clamp m_deathExpireTime.
+#define DEATH_EXPIRE_STEP (5*MINUTE)
+#define MAX_DEATH_COUNT 3
+
+/*********************************************************/
+/***                   LOAD SYSTEM                     ***/
+/*********************************************************/
+
+void Player::_LoadDeclinedNames(QueryResult* result)
+{
+    if (!result)
+    {
+        return;
+    }
+
+    delete m_declinedname;
+    m_declinedname = new DeclinedName;
+
+    Field* fields = result->Fetch();
+    for (int i = 0; i < MAX_DECLINED_NAME_CASES; ++i)
+    {
+        m_declinedname->name[i] = fields[i].GetCppString();
+    }
+
+    delete result;
+}
+
+void Player::_LoadArenaTeamInfo(QueryResult* result)
+{
+    // arenateamid, played_week, played_season, personal_rating
+    memset((void*)&m_uint32Values[PLAYER_FIELD_ARENA_TEAM_INFO_1_1], 0, sizeof(uint32) * MAX_ARENA_SLOT * ARENA_TEAM_END);
+    if (!result)
+    {
+        return;
+    }
+
+    do
+    {
+        Field* fields = result->Fetch();
+
+        uint32 arenateamid     = fields[0].GetUInt32();
+        uint32 played_week     = fields[1].GetUInt32();
+        uint32 played_season   = fields[2].GetUInt32();
+        uint32 wons_season     = fields[3].GetUInt32();
+        uint32 personal_rating = fields[4].GetUInt32();
+
+        ArenaTeam* aTeam = sObjectMgr.GetArenaTeamById(arenateamid);
+        if (!aTeam)
+        {
+            sLog.outError("Player::_LoadArenaTeamInfo: couldn't load arenateam %u", arenateamid);
+            continue;
+        }
+        uint8  arenaSlot = aTeam->GetSlot();
+
+        SetArenaTeamInfoField(arenaSlot, ARENA_TEAM_ID, arenateamid);
+        SetArenaTeamInfoField(arenaSlot, ARENA_TEAM_TYPE, aTeam->GetType());
+        SetArenaTeamInfoField(arenaSlot, ARENA_TEAM_MEMBER, (aTeam->GetCaptainGuid() == GetObjectGuid()) ? 0 : 1);
+        SetArenaTeamInfoField(arenaSlot, ARENA_TEAM_GAMES_WEEK, played_week);
+        SetArenaTeamInfoField(arenaSlot, ARENA_TEAM_GAMES_SEASON, played_season);
+        SetArenaTeamInfoField(arenaSlot, ARENA_TEAM_WINS_SEASON, wons_season);
+        SetArenaTeamInfoField(arenaSlot, ARENA_TEAM_PERSONAL_RATING, personal_rating);
+    }
+    while (result->NextRow());
+    delete result;
+}
+
+void Player::_LoadEquipmentSets(QueryResult* result)
+{
+    // SetPQuery(PLAYER_LOGIN_QUERY_LOADEQUIPMENTSETS,   "SELECT setguid, setindex, name, iconname, ignore_mask, item0, item1, item2, item3, item4, item5, item6, item7, item8, item9, item10, item11, item12, item13, item14, item15, item16, item17, item18 FROM character_equipmentsets WHERE guid = '%u' ORDER BY setindex", GUID_LOPART(m_guid));
+    if (!result)
+    {
+        return;
+    }
+
+    uint32 count = 0;
+    do
+    {
+        Field* fields = result->Fetch();
+
+        EquipmentSet eqSet;
+
+        eqSet.Guid          = fields[0].GetUInt64();
+        uint32 index        = fields[1].GetUInt32();
+        eqSet.Name          = fields[2].GetCppString();
+        eqSet.IconName      = fields[3].GetCppString();
+        eqSet.IgnoreMask    = fields[4].GetUInt32();
+        eqSet.state         = EQUIPMENT_SET_UNCHANGED;
+
+        for (uint32 i = 0; i < EQUIPMENT_SLOT_END; ++i)
+        {
+            eqSet.Items[i] = fields[5 + i].GetUInt32();
+        }
+
+        m_EquipmentSets[index] = eqSet;
+
+        ++count;
+
+        if (count >= MAX_EQUIPMENT_SET_INDEX)               // client limit
+        {
+            break;
+        }
+    }
+    while (result->NextRow());
+    delete result;
+}
+
+/**
+ * @brief Loads battleground return and participation data from the database.
+ *
+ * @param result The query result containing battleground data.
+ */
+void Player::_LoadBGData(QueryResult* result)
+{
+    if (!result)
+    {
+        return;
+    }
+
+    // Expecting only one row
+    Field* fields = result->Fetch();
+    /* bgInstanceID, bgTeam, x, y, z, o, map, taxi[0], taxi[1], mountSpell */
+    m_bgData.bgInstanceID = fields[0].GetUInt32();
+    m_bgData.bgTeam       = Team(fields[1].GetUInt32());
+    m_bgData.joinPos      = WorldLocation(fields[6].GetUInt32(),    // Map
+                                          fields[2].GetFloat(),     // X
+                                          fields[3].GetFloat(),     // Y
+                                          fields[4].GetFloat(),     // Z
+                                          fields[5].GetFloat());    // Orientation
+    m_bgData.taxiPath[0]  = fields[7].GetUInt32();
+    m_bgData.taxiPath[1]  = fields[8].GetUInt32();
+    m_bgData.mountSpell   = fields[9].GetUInt32();
+
+    delete result;
+}
+
+/**
+ * @brief Loads a character position directly from the database.
+ *
+ * @param guid The player GUID to query.
+ * @param mapid Output map identifier.
+ * @param x Output X coordinate.
+ * @param y Output Y coordinate.
+ * @param z Output Z coordinate.
+ * @param o Output orientation.
+ * @param in_flight Output flag indicating whether the character was in flight.
+ * @return True if the position was loaded successfully; otherwise, false.
+ */
+bool Player::LoadPositionFromDB(ObjectGuid guid, uint32& mapid, float& x, float& y, float& z, float& o, bool& in_flight)
+{
+    QueryResult* result = CharacterDatabase.PQuery("SELECT `position_x`,`position_y`,`position_z`,`orientation`,`map`,`taxi_path` FROM `characters` WHERE `guid` = '%u'", guid.GetCounter());
+    if (!result)
+    {
+        return false;
+    }
+
+    Field* fields = result->Fetch();
+
+    x = fields[0].GetFloat();
+    y = fields[1].GetFloat();
+    z = fields[2].GetFloat();
+    o = fields[3].GetFloat();
+    mapid = fields[4].GetUInt32();
+    in_flight = !fields[5].GetCppString().empty();
+
+    delete result;
+    return true;
+}
+
+/**
+ * @brief Loads serialized uint32 values into the player's update field array.
+ *
+ * @param data The serialized space-separated field data.
+ * @param startOffset The first update-field offset to populate.
+ * @param count The number of uint32 values expected.
+ */
+void Player::_LoadIntoDataField(const char* data, uint32 startOffset, uint32 count)
+{
+    if (!data)
+    {
+        return;
+    }
+
+    Tokens tokens = StrSplit(data, " ");
+
+    if (tokens.size() != count)
+    {
+        return;
+    }
+
+    Tokens::iterator iter;
+    uint32 index;
+    for (iter = tokens.begin(), index = 0; index < count; ++iter, ++index)
+    {
+        m_uint32Values[startOffset + index] = std::stoul((*iter).c_str());
+    }
+}
+
+/**
+ * @brief Loads the player's persisted state from the database and initializes runtime data.
+ *
+ * @param guid The player GUID to load.
+ * @param holder The query holder containing login query results.
+ * @return True if the player was loaded successfully; otherwise, false.
+ */
+bool Player::LoadFromDB(ObjectGuid guid, SqlQueryHolder* holder)
+{
+
+    //         0      1         2      3      4       5        6       7    8       9             10             11
+    // SELECT `guid`,`account`,`name`,`race`,`class`,`gender`,`level`,`xp`,`money`,`playerBytes`,`playerBytes2`,`playerFlags`,
+    //  12           13           14           15    16                   17            18         19       20          21          22          23            24                  25           26
+    // `position_x`,`position_y`,`position_z`,`map`,`dungeon_difficulty`,`orientation`,`taximask`,`online`,`cinematic`,`totaltime`,`leveltime`,`logout_time`,`is_logout_resting`,`rest_bonus`,`resettalents_cost`,
+    //  27                  28              29        30        31        32        33          34            35             36         37     38                  39
+    // `resettalents_time`,`primary_trees`,`trans_x`,`trans_y`,`trans_z`,`trans_o`,`transguid`,`extra_flags`,`stable_slots`,`at_login`,`zone`,`death_expire_time`,`taxi_path`,
+    //  40           41           42               43            44               45
+    // `totalKills`,`todayKills`,`yesterdayKills`,`chosenTitle`,`watchedFaction`,`drunk`,
+    //  46       47       48       49       50       51       52          53           54              55               56            57           58     59                    60                 61           62
+    // `health`,`power1`,`power2`,`power3`,`power4`,`power5`,`specCount`,`activeSpec`,`exploredZones`,`equipmentCache`,`knownTitles`,`actionBars`,`slot`,`deleteInfos_Account`,`deleteInfos_Name`,`deleteDate`,`createdDate`
+
+
+
+    QueryResult *result = holder->GetResult(PLAYER_LOGIN_QUERY_LOADFROM);
+
+    if (!result)
+    {
+        sLog.outError("%s not found in table `characters`, can't load. ", guid.GetString().c_str());
+        return false;
+    }
+
+    Field* fields = result->Fetch();
+
+    uint32 dbAccountId = fields[1].GetUInt32();
+
+    // check if the character's account in the db and the logged in account match.
+    // player should be able to load/delete character only with correct account!
+    if (dbAccountId != GetSession()->GetAccountId())
+    {
+        sLog.outError("%s loading from wrong account (is: %u, should be: %u)",
+                      guid.GetString().c_str(), GetSession()->GetAccountId(), dbAccountId);
+        delete result;
+        return false;
+    }
+
+    Object::_Create(guid.GetCounter(), 0, HIGHGUID_PLAYER);
+
+    m_name = fields[2].GetCppString();
+
+    // check name limitations
+    if (ObjectMgr::CheckPlayerName(m_name) != CHAR_NAME_SUCCESS ||
+            (GetSession()->GetSecurity() == SEC_PLAYER && sObjectMgr.IsReservedName(m_name)))
+    {
+        delete result;
+        CharacterDatabase.PExecute("UPDATE `characters` SET `at_login` = `at_login` | '%u' WHERE `guid` ='%u'",
+                                   uint32(AT_LOGIN_RENAME), guid.GetCounter());
+        return false;
+    }
+
+    // overwrite possible wrong/corrupted guid
+    SetGuidValue(OBJECT_FIELD_GUID, guid);
+
+    // overwrite some data fields
+    SetByteValue(UNIT_FIELD_BYTES_0, 0, fields[3].GetUInt8()); // race
+    SetByteValue(UNIT_FIELD_BYTES_0, 1, fields[4].GetUInt8()); // class
+
+    uint8 gender = fields[5].GetUInt8() & 0x01;
+    SetByteValue(UNIT_FIELD_BYTES_0, 2, gender);            // gender
+
+    SetUInt32Value(UNIT_FIELD_LEVEL, fields[6].GetUInt8());
+    SetUInt32Value(PLAYER_XP, fields[7].GetUInt32());
+
+    _LoadIntoDataField(fields[54].GetString(), PLAYER_EXPLORED_ZONES_1, PLAYER_EXPLORED_ZONES_SIZE);
+    _LoadIntoDataField(fields[56].GetString(), PLAYER__FIELD_KNOWN_TITLES, KNOWN_TITLES_SIZE*2);
+
+    InitDisplayIds();                                       // model, scale and model data
+
+    SetFloatValue(UNIT_FIELD_HOVERHEIGHT, 1.0f);
+
+    // just load criteria/achievement data, safe call before any load, and need, because some spell/item/quest loading
+    // can triggering achievement criteria update that will be lost if this call will later
+    m_achievementMgr.LoadFromDB(holder->GetResult(PLAYER_LOGIN_QUERY_LOADACHIEVEMENTS), holder->GetResult(PLAYER_LOGIN_QUERY_LOADCRITERIAPROGRESS));
+
+    uint64 money = fields[8].GetUInt64();
+    if (money > MAX_MONEY_AMOUNT)
+    {
+        money = MAX_MONEY_AMOUNT;
+    }
+    SetMoney(money);
+
+    SetUInt32Value(PLAYER_BYTES, fields[9].GetUInt32());
+    SetUInt32Value(PLAYER_BYTES_2, fields[10].GetUInt32());
+
+    SetByteValue(PLAYER_BYTES_3, 0, gender);
+    SetByteValue(PLAYER_BYTES_3, 1, fields[45].GetUInt8());
+
+    SetUInt32Value(PLAYER_FLAGS, fields[11].GetUInt32());
+    SetInt32Value(PLAYER_FIELD_WATCHED_FACTION_INDEX, fields[44].GetInt32());
+
+    // Action bars state
+    SetByteValue(PLAYER_FIELD_BYTES, 2, fields[57].GetUInt8());
+
+    m_slot = fields[58].GetUInt8();
+
+    // cleanup inventory related item value fields (its will be filled correctly in _LoadInventory)
+    for (uint8 slot = EQUIPMENT_SLOT_START; slot < EQUIPMENT_SLOT_END; ++slot)
+    {
+        SetGuidValue(PLAYER_FIELD_INV_SLOT_HEAD + (slot * 2), ObjectGuid());
+        SetVisibleItemSlot(slot, NULL);
+
+        delete m_items[slot];
+        m_items[slot] = NULL;
+    }
+
+    DEBUG_FILTER_LOG(LOG_FILTER_PLAYER_STATS, "Load Basic value of player %s is: ", m_name.c_str());
+    outDebugStatsValues();
+
+    // Need to call it to initialize m_team (m_team can be calculated from race)
+    // Other way is to saves m_team into characters table.
+    setFactionForRace(getRace());
+    SetCharm(NULL);
+
+    // load home bind and check in same time class/race pair, it used later for restore broken positions
+    if (!_LoadHomeBind(holder->GetResult(PLAYER_LOGIN_QUERY_LOADHOMEBIND)))
+    {
+        delete result;
+        return false;
+    }
+
+    InitPrimaryProfessions();                               // to max set before any spell loaded
+
+    // init saved position, and fix it later if problematic
+    uint32 transGUID = fields[31].GetUInt32();
+    Relocate(fields[12].GetFloat(), fields[13].GetFloat(), fields[14].GetFloat(), fields[16].GetFloat());
+    SetLocationMapId(fields[15].GetUInt32());
+
+    uint32 difficulty = fields[39].GetUInt32();
+    if (difficulty >= MAX_DUNGEON_DIFFICULTY || getLevel() < LEVELREQUIREMENT_HEROIC)
+    {
+        difficulty = DUNGEON_DIFFICULTY_NORMAL;
+    }
+    SetDungeonDifficulty(Difficulty(difficulty));           // may be changed in _LoadGroup
+
+    _LoadGroup(holder->GetResult(PLAYER_LOGIN_QUERY_LOADGROUP));
+
+    _LoadArenaTeamInfo(holder->GetResult(PLAYER_LOGIN_QUERY_LOADARENAINFO));
+
+    // check arena teams integrity
+    for (uint32 arena_slot = 0; arena_slot < MAX_ARENA_SLOT; ++arena_slot)
+    {
+        uint32 arena_team_id = GetArenaTeamId(arena_slot);
+        if (!arena_team_id)
+        {
+            continue;
+        }
+
+        if (ArenaTeam* at = sObjectMgr.GetArenaTeamById(arena_team_id))
+            if (at->HaveMember(GetObjectGuid()))
+            {
+                continue;
+            }
+
+        // arena team not exist or not member, cleanup fields
+        for (int j = 0; j < ARENA_TEAM_END; ++j)
+        {
+            SetArenaTeamInfoField(arena_slot, ArenaTeamInfoType(j), 0);
+        }
+    }
+
+    SetUInt32Value(PLAYER_FIELD_LIFETIME_HONORABLE_KILLS, fields[40].GetUInt32());
+    SetUInt16Value(PLAYER_FIELD_KILLS, 0, fields[41].GetUInt16()); // today
+    SetUInt16Value(PLAYER_FIELD_KILLS, 1, fields[42].GetUInt16()); // yesterday
+
+    _LoadBoundInstances(holder->GetResult(PLAYER_LOGIN_QUERY_LOADBOUNDINSTANCES));
+
+    if (!IsPositionValid())
+    {
+        sLog.outError("%s have invalid coordinates (X: %f Y: %f Z: %f O: %f). Teleport to default race/class locations.",
+                      guid.GetString().c_str(), GetPositionX(), GetPositionY(), GetPositionZ(), GetOrientation());
+        RelocateToHomebind();
+
+        transGUID = 0;
+
+        m_movementInfo.ClearTransportData();
+    }
+
+    _LoadBGData(holder->GetResult(PLAYER_LOGIN_QUERY_LOADBGDATA));
+
+    if (m_bgData.bgInstanceID)                              // saved in BattleGround
+    {
+        BattleGround* currentBg = sBattleGroundMgr.GetBattleGround(m_bgData.bgInstanceID, BATTLEGROUND_TYPE_NONE);
+
+        bool player_at_bg = currentBg && currentBg->IsPlayerInBattleGround(GetObjectGuid());
+
+        if (player_at_bg && currentBg->GetStatus() != STATUS_WAIT_LEAVE)
+        {
+            BattleGroundQueueTypeId bgQueueTypeId = BattleGroundMgr::BGQueueTypeId(currentBg->GetTypeID(), currentBg->GetArenaType());
+            AddBattleGroundQueueId(bgQueueTypeId);
+
+            m_bgData.bgTypeID = currentBg->GetTypeID();     // bg data not marked as modified
+
+            // join player to battleground group
+            currentBg->EventPlayerLoggedIn(this);
+            currentBg->AddOrSetPlayerToCorrectBgGroup(this, GetObjectGuid(), m_bgData.bgTeam);
+
+            SetInviteForBattleGroundQueueType(bgQueueTypeId, currentBg->GetInstanceID());
+        }
+        else
+        {
+            // leave bg
+            if (player_at_bg)
+            {
+                currentBg->RemovePlayerAtLeave(GetObjectGuid(), false, true);
+            }
+
+            // move to bg enter point
+            const WorldLocation& _loc = GetBattleGroundEntryPoint();
+            SetLocationMapId(_loc.mapid);
+            Relocate(_loc.coord_x, _loc.coord_y, _loc.coord_z, _loc.orientation);
+
+            // We are not in BG anymore
+            SetBattleGroundId(0, BATTLEGROUND_TYPE_NONE);
+            // remove outdated DB data in DB
+            _SaveBGData();
+        }
+    }
+    else
+    {
+        MapEntry const* mapEntry = sMapStore.LookupEntry(GetMapId());
+        // if server restart after player save in BG or area
+        // player can have current coordinates in to BG/Arena map, fix this
+        if (!mapEntry || mapEntry->IsBattleGroundOrArena())
+        {
+            const WorldLocation& _loc = GetBattleGroundEntryPoint();
+            SetLocationMapId(_loc.mapid);
+            Relocate(_loc.coord_x, _loc.coord_y, _loc.coord_z, _loc.orientation);
+
+            // We are not in BG anymore
+            SetBattleGroundId(0, BATTLEGROUND_TYPE_NONE);
+            // remove outdated DB data in DB
+            _SaveBGData();
+        }
+    }
+
+    if (transGUID != 0)
+    {
+        m_movementInfo.SetTransportData(ObjectGuid(HIGHGUID_MO_TRANSPORT, transGUID), fields[27].GetFloat(), fields[28].GetFloat(), fields[29].GetFloat(), fields[30].GetFloat(), 0, -1);
+
+        if (!MaNGOS::IsValidMapCoord(
+                    GetPositionX() + m_movementInfo.GetTransportPos()->x, GetPositionY() + m_movementInfo.GetTransportPos()->y,
+                    GetPositionZ() + m_movementInfo.GetTransportPos()->z, GetOrientation() + m_movementInfo.GetTransportPos()->o) ||
+                // transport size limited
+                m_movementInfo.GetTransportPos()->x > 50 || m_movementInfo.GetTransportPos()->y > 50 || m_movementInfo.GetTransportPos()->z > 50)
+        {
+            sLog.outError("%s have invalid transport coordinates (X: %f Y: %f Z: %f O: %f). Teleport to default race/class locations.",
+                          guid.GetString().c_str(), GetPositionX() + m_movementInfo.GetTransportPos()->x, GetPositionY() + m_movementInfo.GetTransportPos()->y,
+                          GetPositionZ() + m_movementInfo.GetTransportPos()->z, GetOrientation() + m_movementInfo.GetTransportPos()->o);
+
+            RelocateToHomebind();
+
+            m_movementInfo.ClearTransportData();
+
+            transGUID = 0;
+        }
+    }
+
+    if (transGUID != 0)
+    {
+        for (MapManager::TransportSet::const_iterator iter = sMapMgr.m_Transports.begin(); iter != sMapMgr.m_Transports.end(); ++iter)
+        {
+            if ((*iter)->GetGUIDLow() == transGUID)
+            {
+                MapEntry const* transMapEntry = sMapStore.LookupEntry((*iter)->GetMapId());
+                // client without expansion support
+                if (GetSession()->Expansion() < transMapEntry->Expansion())
+                {
+                    DEBUG_LOG("Player %s using client without required expansion tried login at transport at non accessible map %u", GetName(), (*iter)->GetMapId());
+                    break;
+                }
+
+                m_transport = *iter;
+                m_transport->AddPassenger(this);
+                SetLocationMapId(m_transport->GetMapId());
+                break;
+            }
+        }
+
+        if (!m_transport)
+        {
+            sLog.outError("%s have problems with transport guid (%u). Teleport to default race/class locations.",
+                          guid.GetString().c_str(), transGUID);
+
+            RelocateToHomebind();
+
+            m_movementInfo.ClearTransportData();
+
+            transGUID = 0;
+        }
+    }
+    else                                                    // not transport case
+    {
+        MapEntry const* mapEntry = sMapStore.LookupEntry(GetMapId());
+        // client without expansion support
+        if (GetSession()->Expansion() < mapEntry->Expansion())
+        {
+            DEBUG_LOG("Player %s using client without required expansion tried login at non accessible map %u", GetName(), GetMapId());
+            RelocateToHomebind();
+        }
+    }
+
+    // player bounded instance saves loaded in _LoadBoundInstances, group versions at group loading
+    DungeonPersistentState* state = GetBoundInstanceSaveForSelfOrGroup(GetMapId());
+
+    // load the player's map here if it's not already loaded
+    SetMap(sMapMgr.CreateMap(GetMapId(), this));
+
+    // if the player is in an instance and it has been reset in the meantime teleport him to the entrance
+    if (GetInstanceId() && !state)
+    {
+        AreaTrigger const* at = sObjectMgr.GetMapEntranceTrigger(GetMapId());
+        if (at)
+        {
+            Relocate(at->target_X, at->target_Y, at->target_Z, at->target_Orientation);
+        }
+        else
+        {
+            sLog.outError("Player %s(GUID: %u) logged in to a reset instance (map: %u) and there is no area-trigger leading to this map. Thus he can't be ported back to the entrance. This _might_ be an exploit attempt.", GetName(), GetGUIDLow(), GetMapId());
+        }
+    }
+
+    SaveRecallPosition();
+
+    time_t now = time(NULL);
+    time_t logoutTime = time_t(fields[22].GetUInt64());
+
+    // since last logout (in seconds)
+    uint32 time_diff = uint32(now - logoutTime);
+
+    // set value, including drunk invisibility detection
+    // calculate sobering. after 15 minutes logged out, the player will be sober again
+    uint8 newDrunkValue = 0;
+    if (time_diff < uint32(GetDrunkValue()) * 9)
+    {
+        newDrunkValue = GetDrunkValue() - time_diff / 9;
+    }
+
+    SetDrunkValue(newDrunkValue);
+
+    m_cinematic = fields[18].GetUInt32();
+    m_Played_time[PLAYED_TIME_TOTAL] = fields[19].GetUInt32();
+    m_Played_time[PLAYED_TIME_LEVEL] = fields[20].GetUInt32();
+
+    m_resetTalentsCost = fields[24].GetUInt32();
+    m_resetTalentsTime = time_t(fields[25].GetUInt64());
+
+    // reserve some flags
+    uint32 old_safe_flags = GetUInt32Value(PLAYER_FLAGS) & (PLAYER_FLAGS_HIDE_CLOAK | PLAYER_FLAGS_HIDE_HELM);
+
+    if (HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_GM))
+    {
+        SetUInt32Value(PLAYER_FLAGS, 0 | old_safe_flags);
+    }
+
+    m_taxi.LoadTaxiMask(fields[17].GetString());            // must be before InitTaxiNodesForLevel
+
+    uint32 extraflags = fields[32].GetUInt32();
+
+    m_petMgr.LoadStableSlotsFromField(fields[33].GetUInt32());
+
+    m_atLoginFlags = fields[34].GetUInt32();
+
+    // Honor system
+    // Update Honor kills data
+    m_honorMgr.SetLastKillUpdate(logoutTime);
+    UpdateHonorKills();
+
+    m_deathExpireTime = (time_t)fields[37].GetUInt64();
+    if (m_deathExpireTime > now + MAX_DEATH_COUNT * DEATH_EXPIRE_STEP)
+    {
+        m_deathExpireTime = now + MAX_DEATH_COUNT * DEATH_EXPIRE_STEP - 1;
+    }
+
+    std::string taxi_nodes = fields[38].GetCppString();
+
+    // clear channel spell data (if saved at channel spell casting)
+    SetChannelObjectGuid(ObjectGuid());
+    SetUInt32Value(UNIT_CHANNEL_SPELL, 0);
+
+    // clear charm/summon related fields
+    SetCharm(NULL);
+    SetPet(NULL);
+    SetTargetGuid(ObjectGuid());
+    SetCharmerGuid(ObjectGuid());
+    SetOwnerGuid(ObjectGuid());
+    SetCreatorGuid(ObjectGuid());
+
+    // reset some aura modifiers before aura apply
+
+    SetGuidValue(PLAYER_FARSIGHT, ObjectGuid());
+    SetUInt32Value(PLAYER_TRACK_CREATURES, 0);
+    SetUInt32Value(PLAYER_TRACK_RESOURCES, 0);
+
+    // cleanup aura list explicitly before skill load where some spells can be applied
+    RemoveAllAuras();
+
+    // make sure the unit is considered out of combat for proper loading
+    ClearInCombat();
+
+    // make sure the unit is considered not in duel for proper loading
+    SetGuidValue(PLAYER_DUEL_ARBITER, ObjectGuid());
+    SetUInt32Value(PLAYER_DUEL_TEAM, 0);
+
+    m_specsCount = fields[52].GetUInt8();
+    m_activeSpec = fields[53].GetUInt8();
+
+    Tokens talentTrees = StrSplit(fields[26].GetString(), " ");
+    for (uint8 i = 0; i < MAX_TALENT_SPEC_COUNT; ++i)
+    {
+        if (i >= talentTrees.size())
+        {
+            break;
+        }
+
+        uint32 talentTree = atol(talentTrees[i].c_str());
+        if (!talentTree || sTalentTabStore.LookupEntry(talentTree))
+        {
+            m_talentsPrimaryTree[i] = talentTree;
+        }
+        else if (i == m_activeSpec)
+        {
+            SetAtLoginFlag(AT_LOGIN_RESET_TALENTS); // invalid tree, reset talents
+        }
+    }
+
+    // reset stats before loading any modifiers
+    InitStatsForLevel();
+    InitGlyphsForLevel();
+    InitTaxiNodesForLevel();
+    InitRunes();
+
+    // rest bonus can only be calculated after InitStatsForLevel()
+    m_rest_bonus = fields[21].GetFloat();
+
+    if (time_diff > 0)
+    {
+        SetRestBonus(GetRestBonus() + ComputeRest(time_diff, true, (fields[23].GetInt32() > 0)));
+    }
+
+    // load skills after InitStatsForLevel because it triggering aura apply also
+    _LoadSkills(holder->GetResult(PLAYER_LOGIN_QUERY_LOADSKILLS));
+
+    // apply original stats mods before spell loading or item equipment that call before equip _RemoveStatsMods()
+
+    // Mail
+    _LoadMails(holder->GetResult(PLAYER_LOGIN_QUERY_LOADMAILS));
+    _LoadMailedItems(holder->GetResult(PLAYER_LOGIN_QUERY_LOADMAILEDITEMS));
+    UpdateNextMailTimeAndUnreads();
+
+    _LoadGlyphs(holder->GetResult(PLAYER_LOGIN_QUERY_LOADGLYPHS));
+
+    _LoadAuras(holder->GetResult(PLAYER_LOGIN_QUERY_LOADAURAS), time_diff);
+    ApplyGlyphs(true);
+
+    // add ghost flag (must be after aura load: PLAYER_FLAGS_GHOST set in aura)
+    if (HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_GHOST))
+    {
+        m_deathState = DEAD;
+    }
+
+    _LoadCurrencies(holder->GetResult(PLAYER_LOGIN_QUERY_LOADCURRENCIES));
+    _LoadSpells(holder->GetResult(PLAYER_LOGIN_QUERY_LOADSPELLS));
+
+    // after spell load, learn rewarded spell if need also
+    _LoadQuestStatus(holder->GetResult(PLAYER_LOGIN_QUERY_LOADQUESTSTATUS));
+    _LoadDailyQuestStatus(holder->GetResult(PLAYER_LOGIN_QUERY_LOADDAILYQUESTSTATUS));
+    _LoadWeeklyQuestStatus(holder->GetResult(PLAYER_LOGIN_QUERY_LOADWEEKLYQUESTSTATUS));
+    _LoadMonthlyQuestStatus(holder->GetResult(PLAYER_LOGIN_QUERY_LOADMONTHLYQUESTSTATUS));
+
+    _LoadTalents(holder->GetResult(PLAYER_LOGIN_QUERY_LOADTALENTS));
+
+    // after spell and quest load
+    InitTalentForLevel();
+    learnDefaultSpells();
+
+    // must be before inventory (some items required reputation check)
+    m_reputationMgr.LoadFromDB(holder->GetResult(PLAYER_LOGIN_QUERY_LOADREPUTATION));
+
+    _LoadInventory(holder->GetResult(PLAYER_LOGIN_QUERY_LOADINVENTORY), time_diff);
+    _LoadItemLoot(holder->GetResult(PLAYER_LOGIN_QUERY_LOADITEMLOOT));
+
+    // update items with duration and realtime
+    UpdateItemDuration(time_diff, true);
+
+    _LoadActions(holder->GetResult(PLAYER_LOGIN_QUERY_LOADACTIONS));
+
+    m_social = sSocialMgr.LoadFromDB(holder->GetResult(PLAYER_LOGIN_QUERY_LOADSOCIALLIST), GetObjectGuid());
+
+    // check PLAYER_CHOSEN_TITLE compatibility with PLAYER__FIELD_KNOWN_TITLES
+    // note: PLAYER__FIELD_KNOWN_TITLES updated at quest status loaded
+    uint32 curTitle = fields[43].GetUInt32();
+    if (curTitle && !HasTitle(curTitle))
+    {
+        curTitle = 0;
+    }
+
+    SetUInt32Value(PLAYER_CHOSEN_TITLE, curTitle);
+
+    // Not finish taxi flight path
+    if (m_bgData.HasTaxiPath())
+    {
+        m_taxi.ClearTaxiDestinations();
+        for (int i = 0; i < 2; ++i)
+        {
+            m_taxi.AddTaxiDestination(m_bgData.taxiPath[i]);
+        }
+    }
+    else if (!m_taxi.LoadTaxiDestinationsFromString(taxi_nodes, GetTeam()))
+    {
+        // problems with taxi path loading
+        TaxiNodesEntry const* nodeEntry = NULL;
+        if (uint32 node_id = m_taxi.GetTaxiSource())
+        {
+            nodeEntry = sTaxiNodesStore.LookupEntry(node_id);
+        }
+
+        if (!nodeEntry)                                     // don't know taxi start node, to homebind
+        {
+            sLog.outError("Character %u have wrong data in taxi destination list, teleport to homebind.", GetGUIDLow());
+            RelocateToHomebind();
+        }
+        else                                                // have start node, to it
+        {
+            sLog.outError("Character %u have too short taxi destination list, teleport to original node.", GetGUIDLow());
+            SetLocationMapId(nodeEntry->map_id);
+            Relocate(nodeEntry->x, nodeEntry->y, nodeEntry->z, 0.0f);
+        }
+
+        // we can be relocated from taxi and still have an outdated Map pointer!
+        // so we need to get a new Map pointer!
+        SetMap(sMapMgr.CreateMap(GetMapId(), this));
+        SaveRecallPosition();                           // save as recall also to prevent recall and fall from sky
+
+        m_taxi.ClearTaxiDestinations();
+    }
+
+    if (uint32 node_id = m_taxi.GetTaxiSource())
+    {
+        // save source node as recall coord to prevent recall and fall from sky
+        TaxiNodesEntry const* nodeEntry = sTaxiNodesStore.LookupEntry(node_id);
+        MANGOS_ASSERT(nodeEntry);                           // checked in m_taxi.LoadTaxiDestinationsFromString
+        m_recallMap = nodeEntry->map_id;
+        m_recallX = nodeEntry->x;
+        m_recallY = nodeEntry->y;
+        m_recallZ = nodeEntry->z;
+
+        // flight will started later
+    }
+
+    // has to be called after last Relocate() in Player::LoadFromDB
+    SetFallInformation(0, GetPositionZ());
+
+    _LoadSpellCooldowns(holder->GetResult(PLAYER_LOGIN_QUERY_LOADSPELLCOOLDOWNS));
+
+    // Spell code allow apply any auras to dead character in load time in aura/spell/item loading
+    // Do now before stats re-calculation cleanup for ghost state unexpected auras
+    if (!IsAlive())
+    {
+        RemoveAllAurasOnDeath();
+    }
+
+    // apply all stat bonuses from items and auras
+    SetCanModifyStats(true);
+    UpdateAllStats();
+
+    // restore remembered power/health values (but not more max values)
+    uint32 savedhealth = fields[46].GetUInt32();
+    SetHealth(savedhealth > GetMaxHealth() ? GetMaxHealth() : savedhealth);
+    for (uint32 i = 0; i < MAX_STORED_POWERS; ++i)
+    {
+        uint32 savedpower = fields[47 + i].GetUInt32();
+        SetPowerByIndex(i, std::min(savedpower, GetMaxPowerByIndex(i)));
+    }
+
+    uint32 createdDate = fields[59].GetUInt32();
+    SetCreatedDate(createdDate);
+
+    DEBUG_FILTER_LOG(LOG_FILTER_PLAYER_STATS, "The value of player %s after load item and aura is: ", m_name.c_str());
+    outDebugStatsValues();
+
+    // all fields read
+    delete result;
+
+    // GM state
+    if (GetSession()->GetSecurity() > SEC_PLAYER)
+    {
+        switch (sWorld.getConfig(CONFIG_UINT32_GM_LOGIN_STATE))
+        {
+            default:
+            case 0:                      break;             // disable
+            case 1: SetGameMaster(true); break;             // enable
+            case 2:                                         // save state
+                if (extraflags & PLAYER_EXTRA_GM_ON)
+                {
+                    SetGameMaster(true);
+                }
+                break;
+        }
+
+        switch (sWorld.getConfig(CONFIG_UINT32_GM_VISIBLE_STATE))
+        {
+            default:
+            case 0: SetGMVisible(false); break;             // invisible
+            case 1:                      break;             // visible
+            case 2:                                         // save state
+                if (extraflags & PLAYER_EXTRA_GM_INVISIBLE)
+                {
+                    SetGMVisible(false);
+                }
+                break;
+        }
+
+        switch (sWorld.getConfig(CONFIG_UINT32_GM_ACCEPT_TICKETS))
+        {
+            default:
+            case 0:                        break;           // disable
+            case 1: SetAcceptTicket(true); break;           // enable
+            case 2:                                         // save state
+                if (extraflags & PLAYER_EXTRA_GM_ACCEPT_TICKETS)
+                {
+                    SetAcceptTicket(true);
+                }
+                break;
+        }
+
+        switch (sWorld.getConfig(CONFIG_UINT32_GM_CHAT))
+        {
+            default:
+            case 0:                  break;                 // disable
+            case 1: SetGMChat(true); break;                 // enable
+            case 2:                                         // save state
+                if (extraflags & PLAYER_EXTRA_GM_CHAT)
+                {
+                    SetGMChat(true);
+                }
+                break;
+        }
+
+        switch (sWorld.getConfig(CONFIG_UINT32_GM_WISPERING_TO))
+        {
+            default:
+            case 0:                          break;         // disable
+            case 1: SetAcceptWhispers(true); break;         // enable
+            case 2:                                         // save state
+                if (extraflags & PLAYER_EXTRA_ACCEPT_WHISPERS)
+                {
+                    SetAcceptWhispers(true);
+                }
+                break;
+        }
+    }
+
+    _LoadDeclinedNames(holder->GetResult(PLAYER_LOGIN_QUERY_LOADDECLINEDNAMES));
+
+    m_achievementMgr.CheckAllAchievementCriteria();
+
+    _LoadEquipmentSets(holder->GetResult(PLAYER_LOGIN_QUERY_LOADEQUIPMENTSETS));
+
+    return true;
+}
+
+
+/**
+ * @brief Loads action bar bindings from the database.
+ *
+ * @param result The query result containing action bindings.
+ */
+void Player::_LoadActions(QueryResult* result)
+{
+    for (int i = 0; i < MAX_TALENT_SPEC_COUNT; ++i)
+    {
+        m_actionButtons[i].clear();
+    }
+
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `spec`, `button`,`action`,`type` FROM `character_action` WHERE `guid` = '%u' ORDER BY `button`",GetGUIDLow());
+
+    if (result)
+    {
+        do
+        {
+            Field* fields = result->Fetch();
+
+            uint8 spec = fields[0].GetUInt8();
+            uint8 button = fields[1].GetUInt8();
+            uint32 action = fields[2].GetUInt32();
+            uint8 type = fields[3].GetUInt8();
+
+            if (ActionButton* ab = addActionButton(spec, button, action, type))
+            {
+                ab->uState = ACTIONBUTTON_UNCHANGED;
+            }
+            else
+            {
+                sLog.outError("  ...at loading, and will deleted in DB also");
+
+                // Will deleted in DB at next save (it can create data until save but marked as deleted)
+                m_actionButtons[spec][button].uState = ACTIONBUTTON_DELETED;
+            }
+        }
+        while (result->NextRow());
+
+        delete result;
+    }
+}
+
+/**
+ * @brief Loads saved auras from the database and restores their remaining durations.
+ *
+ * @param result The query result containing aura data.
+ * @param timediff The elapsed offline time used to age aura durations.
+ */
+void Player::_LoadAuras(QueryResult* result, uint32 timediff)
+{
+    // RemoveAllAuras(); -- some spells casted before aura load, for example in LoadSkills, aura list explicitly cleaned early
+
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `caster_guid`,`item_guid`,`spell`,`stackcount`,`remaincharges`,`basepoints0`,`basepoints1`,`basepoints2`,`periodictime0`,`periodictime1`,`periodictime2`,`maxduration`,`remaintime`,`effIndexMask` FROM `character_aura` WHERE `guid` = '%u'",GetGUIDLow());
+
+    if (result)
+    {
+        do
+        {
+            Field* fields = result->Fetch();
+            ObjectGuid caster_guid = ObjectGuid(fields[0].GetUInt64());
+            uint32 item_lowguid = fields[1].GetUInt32();
+            uint32 spellid = fields[2].GetUInt32();
+            uint32 stackcount = fields[3].GetUInt32();
+            uint32 remaincharges = fields[4].GetUInt32();
+            int32  damage[MAX_EFFECT_INDEX];
+            uint32 periodicTime[MAX_EFFECT_INDEX];
+
+            for (int32 i = 0; i < MAX_EFFECT_INDEX; ++i)
+            {
+                damage[i] = fields[i + 5].GetInt32();
+                periodicTime[i] = fields[i + 8].GetUInt32();
+            }
+
+            int32 maxduration = fields[11].GetInt32();
+            int32 remaintime = fields[12].GetInt32();
+            uint32 effIndexMask = fields[13].GetUInt32();
+
+            SpellEntry const* spellproto = sSpellStore.LookupEntry(spellid);
+            if (!spellproto)
+            {
+                sLog.outError("Unknown spell (spellid %u), ignore.", spellid);
+                continue;
+            }
+
+            if (remaintime != -1 && !IsPositiveSpell(spellproto))
+            {
+                if (remaintime / IN_MILLISECONDS <= int32(timediff))
+                {
+                    continue;
+                }
+
+                remaintime -= timediff * IN_MILLISECONDS;
+            }
+
+            // prevent wrong values of remaincharges
+            if (uint32 procCharges = spellproto->GetProcCharges())
+            {
+                if (remaincharges <= 0 || remaincharges > procCharges)
+                {
+                    remaincharges = procCharges;
+                }
+            }
+            else
+            {
+                remaincharges = 0;
+            }
+
+            uint32 defstackamount = spellproto->GetStackAmount();
+            if (!defstackamount)
+            {
+                stackcount = 1;
+            }
+            else if (defstackamount < stackcount)
+            {
+                stackcount = defstackamount;
+            }
+            else if (!stackcount)
+            {
+                stackcount = 1;
+            }
+
+            SpellAuraHolder* holder = CreateSpellAuraHolder(spellproto, this, NULL);
+            holder->SetLoadedState(caster_guid, ObjectGuid(HIGHGUID_ITEM, item_lowguid), stackcount, remaincharges, maxduration, remaintime);
+
+            for (int32 i = 0; i < MAX_EFFECT_INDEX; ++i)
+            {
+                if ((effIndexMask & (1 << i)) == 0)
+                {
+                    continue;
+                }
+
+                Aura* aura = CreateAura(spellproto, SpellEffectIndex(i), NULL, holder, this);
+                if (!damage[i])
+                {
+                    damage[i] = aura->GetModifier()->m_amount;
+                }
+
+                aura->SetLoadedState(damage[i], periodicTime[i]);
+                holder->AddAura(aura, SpellEffectIndex(i));
+            }
+
+            if (!holder->IsEmptyHolder())
+            {
+                // reset stolen single target auras
+                if (caster_guid != GetObjectGuid() && holder->GetTrackedAuraType() == TRACK_AURA_TYPE_SINGLE_TARGET)
+                {
+                    holder->SetTrackedAuraType(TRACK_AURA_TYPE_NOT_TRACKED);
+                }
+
+                AddSpellAuraHolder(holder);
+                DETAIL_LOG("Added auras from spellid %u", spellproto->Id);
+            }
+            else
+            {
+                delete holder;
+            }
+        }
+        while (result->NextRow());
+        delete result;
+    }
+
+    if (getClass() == CLASS_WARRIOR && !HasAuraType(SPELL_AURA_MOD_SHAPESHIFT))
+    {
+        CastSpell(this, SPELL_ID_PASSIVE_BATTLE_STANCE, true);
+    }
+}
+
+// Player::_LoadGlyphs moved to GlyphMgr::Load (2026-05-12); thin delegating wrapper lives inline in Player.h.
+
+/**
+ * @brief Restores corpse state for a dead player or cleans it up for a living one.
+ */
+void Player::LoadCorpse()
+{
+    if (IsAlive())
+    {
+        sObjectAccessor.ConvertCorpseForPlayer(GetObjectGuid());
+    }
+    else
+    {
+        if (Corpse* corpse = GetCorpse())
+        {
+            ApplyModByteFlag(PLAYER_FIELD_BYTES, 0, PLAYER_FIELD_BYTE_RELEASE_TIMER, corpse && !sMapStore.LookupEntry(corpse->GetMapId())->Instanceable());
+        }
+        else
+        {
+            // Prevent Dead Player login without corpse
+            ResurrectPlayer(0.5f);
+        }
+    }
+}
+
+/**
+ * @brief Loads inventory, bank, and equipped items from the database.
+ *
+ * @param result The query result containing item inventory rows.
+ * @param timediff The elapsed offline time used for time-sensitive item checks.
+ */
+void Player::_LoadInventory(QueryResult* result, uint32 timediff)
+{
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `data`,`text`,`bag`,`slot`,`item`,`item_template` FROM `character_inventory` JOIN `item_instance` ON `character_inventory`.`item` = `item_instance`.`guid` WHERE `character_inventory`.`guid` = '%u' ORDER BY `bag`,`slot`", GetGUIDLow());
+    std::map<uint32, Bag*> bagMap;                          // fast guid lookup for bags
+    // NOTE: the "order by `bag`" is important because it makes sure
+    // the bagMap is filled before items in the bags are loaded
+    // NOTE2: the "order by `slot`" is needed because mainhand weapons are (wrongly?)
+    // expected to be equipped before offhand items (TODO: fixme)
+
+    uint32 zone = GetZoneId();
+
+    if (result)
+    {
+        std::list<Item*> problematicItems;
+
+        // prevent items from being added to the queue when stored
+        m_itemUpdateQueueBlocked = true;
+        do
+        {
+            Field* fields = result->Fetch();
+            uint32 bag_guid  = fields[2].GetUInt32();
+            uint8  slot      = fields[3].GetUInt8();
+            uint32 item_lowguid = fields[4].GetUInt32();
+            uint32 item_id   = fields[5].GetUInt32();
+
+            ItemPrototype const* proto = ObjectMgr::GetItemPrototype(item_id);
+
+            if (!proto)
+            {
+                CharacterDatabase.PExecute("DELETE FROM `character_inventory` WHERE `item` = '%u'", item_lowguid);
+                CharacterDatabase.PExecute("DELETE FROM `item_instance` WHERE `guid` = '%u'", item_lowguid);
+                sLog.outError("Player::_LoadInventory: Player %s has an unknown item (id: #%u) in inventory, deleted.", GetName(), item_id);
+                continue;
+            }
+
+            Item* item = NewItemOrBag(proto);
+
+            if (!item->LoadFromDB(item_lowguid, fields, GetObjectGuid()))
+            {
+                sLog.outError("Player::_LoadInventory: Player %s has broken item (id: #%u) in inventory, deleted.", GetName(), item_id);
+                CharacterDatabase.PExecute("DELETE FROM `character_inventory` WHERE `item` = '%u'", item_lowguid);
+                item->FSetState(ITEM_REMOVED);
+                item->SaveToDB();                           // it also deletes item object !
+                continue;
+            }
+
+            // not allow have in alive state item limited to another map/zone
+            if (IsAlive() && item->IsLimitedToAnotherMapOrZone(GetMapId(), zone))
+            {
+                CharacterDatabase.PExecute("DELETE FROM `character_inventory` WHERE `item` = '%u'", item_lowguid);
+                item->FSetState(ITEM_REMOVED);
+                item->SaveToDB();                           // it also deletes item object !
+                continue;
+            }
+
+            // "Conjured items disappear if you are logged out for more than 15 minutes"
+            if (timediff > 15 * MINUTE && (item->GetProto()->Flags & ITEM_FLAG_CONJURED))
+            {
+                CharacterDatabase.PExecute("DELETE FROM `character_inventory` WHERE `item` = '%u'", item_lowguid);
+                item->FSetState(ITEM_REMOVED);
+                item->SaveToDB();                           // it also deletes item object !
+                continue;
+            }
+
+            bool success = true;
+
+            // the item/bag is not in a bag
+            if (!bag_guid)
+            {
+                item->SetContainer(NULL);
+                item->SetSlot(slot);
+
+                if (IsInventoryPos(INVENTORY_SLOT_BAG_0, slot))
+                {
+                    ItemPosCountVec dest;
+                    if (CanStoreItem(INVENTORY_SLOT_BAG_0, slot, dest, item, false) == EQUIP_ERR_OK)
+                    {
+                        item = StoreItem(dest, item, true);
+                    }
+                    else
+                    {
+                        success = false;
+                    }
+                }
+                else if (IsEquipmentPos(INVENTORY_SLOT_BAG_0, slot))
+                {
+                    uint16 dest;
+                    if (CanEquipItem(slot, dest, item, false, false) == EQUIP_ERR_OK)
+                    {
+                        QuickEquipItem(dest, item);
+                    }
+                    else
+                    {
+                        success = false;
+                    }
+                }
+                else if (IsBankPos(INVENTORY_SLOT_BAG_0, slot))
+                {
+                    ItemPosCountVec dest;
+                    if (CanBankItem(INVENTORY_SLOT_BAG_0, slot, dest, item, false, false) == EQUIP_ERR_OK)
+                    {
+                        item = BankItem(dest, item, true);
+                    }
+                    else
+                    {
+                        success = false;
+                    }
+                }
+
+                if (success)
+                {
+                    // store bags that may contain items in them
+                    if (item->IsBag() && IsBagPos(item->GetPos()))
+                    {
+                        bagMap[item_lowguid] = (Bag*)item;
+                    }
+                }
+            }
+            // the item/bag in a bag
+            else
+            {
+                item->SetSlot(NULL_SLOT);
+                // the item is in a bag, find the bag
+                std::map<uint32, Bag*>::const_iterator itr = bagMap.find(bag_guid);
+                if (itr != bagMap.end() && slot < itr->second->GetBagSize())
+                {
+                    ItemPosCountVec dest;
+                    if (CanStoreItem(itr->second->GetSlot(), slot, dest, item, false) == EQUIP_ERR_OK)
+                    {
+                        item = StoreItem(dest, item, true);
+                    }
+                    else
+                    {
+                        success = false;
+                    }
+                }
+                else
+                {
+                    success = false;
+                }
+            }
+
+            // item's state may have changed after stored
+            if (success)
+            {
+                item->SetState(ITEM_UNCHANGED, this);
+
+                // restore container unchanged state also
+                if (item->GetContainer())
+                {
+                    item->GetContainer()->SetState(ITEM_UNCHANGED, this);
+                }
+
+                // recharged mana gem
+                if (timediff > 15 * MINUTE && proto->ItemLimitCategory == ITEM_LIMIT_CATEGORY_MANA_GEM)
+                {
+                    item->RestoreCharges();
+                }
+            }
+            else
+            {
+                sLog.outError("Player::_LoadInventory: Player %s has item (GUID: %u Entry: %u) can't be loaded to inventory (Bag GUID: %u Slot: %u) by some reason, will send by mail.", GetName(), item_lowguid, item_id, bag_guid, slot);
+                CharacterDatabase.PExecute("DELETE FROM `character_inventory` WHERE `item` = '%u'", item_lowguid);
+                problematicItems.push_back(item);
+            }
+        }
+        while (result->NextRow());
+
+        delete result;
+        m_itemUpdateQueueBlocked = false;
+
+        // send by mail problematic items
+        while (!problematicItems.empty())
+        {
+            std::string subject = GetSession()->GetMangosString(LANG_NOT_EQUIPPED_ITEM);
+
+            // fill mail
+            MailDraft draft(subject, "There's were problems with equipping item(s).");
+
+            for (int i = 0; !problematicItems.empty() && i < MAX_MAIL_ITEMS; ++i)
+            {
+                Item* item = problematicItems.front();
+                problematicItems.pop_front();
+
+                draft.AddItem(item);
+            }
+
+            draft.SendMailTo(this, MailSender(this, MAIL_STATIONERY_GM), MAIL_CHECK_MASK_COPIED);
+        }
+    }
+
+    // if (IsAlive())
+    _ApplyAllItemMods();
+}
+
+/**
+ * @brief Loads persisted item loot contents for the player's items.
+ *
+ * @param result The query result containing saved item loot rows.
+ */
+void Player::_LoadItemLoot(QueryResult* result)
+{
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `guid`,`itemid`,`amount`,`suffix`,`property` FROM `item_loot` WHERE `guid` = '%u'", GetGUIDLow());
+
+    if (result)
+    {
+        do
+        {
+            Field* fields = result->Fetch();
+            uint32 item_guid   = fields[0].GetUInt32();
+
+            Item* item = GetItemByGuid(ObjectGuid(HIGHGUID_ITEM, item_guid));
+
+            if (!item)
+            {
+                CharacterDatabase.PExecute("DELETE FROM `item_loot` WHERE `guid` = '%u'", item_guid);
+                sLog.outError("Player::_LoadItemLoot: Player %s has loot for nonexistent item (GUID: %u) in `item_loot`, deleted.", GetName(), item_guid);
+                continue;
+            }
+
+            item->LoadLootFromDB(fields);
+        }
+        while (result->NextRow());
+
+        delete result;
+    }
+}
+
+// load mailed item which should receive current player
+void Player::_LoadMailedItems(QueryResult* result)
+{
+    // data needs to be at first place for Item::LoadFromDB
+    //         0     1     2        3          4
+    // "SELECT data, text, mail_id, item_guid, item_template FROM mail_items JOIN item_instance ON item_guid = guid WHERE receiver = '%u'", GUID_LOPART(m_guid)
+    if (!result)
+    {
+        return;
+    }
+
+    do
+    {
+        Field* fields = result->Fetch();
+        uint32 mail_id       = fields[2].GetUInt32();
+        uint32 item_guid_low = fields[3].GetUInt32();
+        uint32 item_template = fields[4].GetUInt32();
+
+        Mail* mail = GetMail(mail_id);
+        if (!mail)
+        {
+            continue;
+        }
+        mail->AddItem(item_guid_low, item_template);
+
+        ItemPrototype const* proto = ObjectMgr::GetItemPrototype(item_template);
+
+        if (!proto)
+        {
+            sLog.outError("Player %u has unknown item_template (ProtoType) in mailed items(GUID: %u template: %u) in mail (%u), deleted.", GetGUIDLow(), item_guid_low, item_template, mail->messageID);
+            CharacterDatabase.PExecute("DELETE FROM `mail_items` WHERE `item_guid` = '%u'", item_guid_low);
+            CharacterDatabase.PExecute("DELETE FROM `item_instance` WHERE `guid` = '%u'", item_guid_low);
+            continue;
+        }
+
+        Item* item = NewItemOrBag(proto);
+
+        if (!item->LoadFromDB(item_guid_low, fields, GetObjectGuid()))
+        {
+            sLog.outError("Player::_LoadMailedItems - Item in mail (%u) doesn't exist !!!! - item guid: %u, deleted from mail", mail->messageID, item_guid_low);
+            CharacterDatabase.PExecute("DELETE FROM `mail_items` WHERE `item_guid` = '%u'", item_guid_low);
+            item->FSetState(ITEM_REMOVED);
+            item->SaveToDB();                               // it also deletes item object !
+            continue;
+        }
+
+        AddMItem(item);
+    }
+    while (result->NextRow());
+
+    delete result;
+}
+
+/**
+ * @brief Loads the player's mailbox headers from the database.
+ *
+ * @param result The query result containing mail records.
+ */
+void Player::_LoadMails(QueryResult* result)
+{
+    m_mail.clear();
+    //        0  1           2      3        4       5    6           7            8     9   10      11         12             13
+    //"SELECT id,messageType,sender,receiver,subject,body,expire_time,deliver_time,money,cod,checked,stationery,mailTemplateId,has_items FROM mail WHERE receiver = '%u' ORDER BY id DESC", GetGUIDLow()
+    if (!result)
+    {
+        return;
+    }
+
+    do
+    {
+        Field* fields = result->Fetch();
+        Mail* m = new Mail;
+        m->messageID = fields[0].GetUInt32();
+        m->messageType = fields[1].GetUInt8();
+        m->sender = fields[2].GetUInt32();
+        m->receiverGuid = ObjectGuid(HIGHGUID_PLAYER, fields[3].GetUInt32());
+        m->subject = fields[4].GetCppString();
+        m->body = fields[5].GetCppString();
+        m->expire_time = (time_t)fields[6].GetUInt64();
+        m->deliver_time = (time_t)fields[7].GetUInt64();
+        m->money = fields[8].GetUInt32();
+        m->COD = fields[9].GetUInt32();
+        m->checked = fields[10].GetUInt32();
+        m->stationery = fields[11].GetUInt8();
+        m->mailTemplateId = fields[12].GetInt16();
+        m->has_items = fields[13].GetBool();                // true, if mail have items or mail have template and items generated (maybe none)
+
+        if (m->mailTemplateId && !sMailTemplateStore.LookupEntry(m->mailTemplateId))
+        {
+            sLog.outError("Player::_LoadMail - Mail (%u) have nonexistent MailTemplateId (%u), remove at load", m->messageID, m->mailTemplateId);
+            m->mailTemplateId = 0;
+        }
+
+        m->state = MAIL_STATE_UNCHANGED;
+
+        m_mail.push_back(m);
+
+        if (m->mailTemplateId && !m->has_items)
+        {
+            m->prepareTemplateItems(this);
+        }
+    }
+    while (result->NextRow());
+    delete result;
+}
+
+/**
+ * @brief Loads the player's currently active pet from the database when possible.
+ */
+void Player::LoadPet()
+{
+    // fixme: the pet should still be loaded if the player is not in world
+    // just not added to the map
+    if (IsInWorld())
+    {
+        Pet* pet = new Pet;
+        if (!pet->LoadPetFromDB(this, 0, 0, true))
+        {
+            delete pet;
+        }
+    }
+}
+
+/**
+ * @brief Loads quest status records and rebuilds the in-memory quest log state.
+ *
+ * @param result The query result containing quest status rows.
+ */
+void Player::_LoadQuestStatus(QueryResult* result)
+{
+    mQuestStatus.clear();
+
+    uint32 slot = 0;
+
+    ////                                                       0        1         2           3           4        5            6            7            8            9             10            11            12            13            14
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `quest`, `status`, `rewarded`, `explored`, `timer`, `mobcount1`, `mobcount2`, `mobcount3`, `mobcount4`, `itemcount1`, `itemcount2`, `itemcount3`, `itemcount4`, `itemcount5`, `itemcount6` FROM `character_queststatus` WHERE `guid` = '%u'", GetGUIDLow());
+
+    if (result)
+    {
+        do
+        {
+            Field* fields = result->Fetch();
+
+            uint32 quest_id = fields[0].GetUInt32();
+            // used to be new, no delete?
+            Quest const* pQuest = sObjectMgr.GetQuestTemplate(quest_id);
+            if (pQuest)
+            {
+                // find or create
+                QuestStatusData& questStatusData = mQuestStatus[quest_id];
+
+                uint32 qstatus = fields[1].GetUInt32();
+                if (qstatus < MAX_QUEST_STATUS)
+                {
+                    questStatusData.m_status = QuestStatus(qstatus);
+                }
+                else
+                {
+                    questStatusData.m_status = QUEST_STATUS_NONE;
+                    sLog.outError("Player %s have invalid quest %d status (%d), replaced by QUEST_STATUS_NONE(0).", GetName(), quest_id, qstatus);
+                }
+
+                questStatusData.m_rewarded = (fields[2].GetUInt8() > 0);
+                questStatusData.m_explored = (fields[3].GetUInt8() > 0);
+
+                time_t quest_time = time_t(fields[4].GetUInt64());
+
+                if (pQuest->HasSpecialFlag(QUEST_SPECIAL_FLAG_TIMED) && !GetQuestRewardStatus(quest_id) && questStatusData.m_status != QUEST_STATUS_NONE)
+                {
+                    AddTimedQuest(quest_id);
+
+                    if (quest_time <= sWorld.GetGameTime())
+                    {
+                        questStatusData.m_timer = 1;
+                    }
+                    else
+                    {
+                        questStatusData.m_timer = uint32(quest_time - sWorld.GetGameTime()) * IN_MILLISECONDS;
+                    }
+                }
+                else
+                {
+                    quest_time = 0;
+                }
+
+                questStatusData.m_creatureOrGOcount[0] = fields[5].GetUInt32();
+                questStatusData.m_creatureOrGOcount[1] = fields[6].GetUInt32();
+                questStatusData.m_creatureOrGOcount[2] = fields[7].GetUInt32();
+                questStatusData.m_creatureOrGOcount[3] = fields[8].GetUInt32();
+                questStatusData.m_itemcount[0] = fields[9].GetUInt32();
+                questStatusData.m_itemcount[1] = fields[10].GetUInt32();
+                questStatusData.m_itemcount[2] = fields[11].GetUInt32();
+                questStatusData.m_itemcount[3] = fields[12].GetUInt32();
+                questStatusData.m_itemcount[4] = fields[13].GetUInt32();
+                questStatusData.m_itemcount[5] = fields[14].GetUInt32();
+
+                questStatusData.uState = QUEST_UNCHANGED;
+
+                // add to quest log
+                if (slot < MAX_QUEST_LOG_SIZE &&
+                        ((questStatusData.m_status == QUEST_STATUS_INCOMPLETE ||
+                          questStatusData.m_status == QUEST_STATUS_COMPLETE ||
+                          questStatusData.m_status == QUEST_STATUS_FAILED) &&
+                         (!questStatusData.m_rewarded || pQuest->IsRepeatable())))
+                {
+                    SetQuestSlot(slot, quest_id, uint32(quest_time));
+
+                    if (questStatusData.m_explored)
+                    {
+                        SetQuestSlotState(slot, QUEST_STATE_COMPLETE);
+                    }
+
+                    if (questStatusData.m_status == QUEST_STATUS_COMPLETE)
+                    {
+                        SetQuestSlotState(slot, QUEST_STATE_COMPLETE);
+                    }
+
+                    if (questStatusData.m_status == QUEST_STATUS_FAILED)
+                    {
+                        SetQuestSlotState(slot, QUEST_STATE_FAIL);
+                    }
+
+                    for (uint8 idx = 0; idx < QUEST_OBJECTIVES_COUNT; ++idx)
+                        if (questStatusData.m_creatureOrGOcount[idx])
+                        {
+                            SetQuestSlotCounter(slot, idx, questStatusData.m_creatureOrGOcount[idx]);
+                        }
+
+                    ++slot;
+                }
+
+                if (questStatusData.m_rewarded)
+                {
+                    // learn rewarded spell if unknown
+                    learnQuestRewardedSpells(pQuest);
+
+                    // set rewarded title if any
+                    if (pQuest->GetCharTitleId())
+                    {
+                        if (CharTitlesEntry const* titleEntry = sCharTitlesStore.LookupEntry(pQuest->GetCharTitleId()))
+                        {
+                            SetTitle(titleEntry);
+                        }
+                    }
+
+                    if (pQuest->GetBonusTalents())
+                    {
+                        m_questRewardTalentCount += pQuest->GetBonusTalents();
+                    }
+                }
+
+                DEBUG_LOG("Quest status is {%u} for quest {%u} for player (GUID: %u)", questStatusData.m_status, quest_id, GetGUIDLow());
+            }
+        }
+        while (result->NextRow());
+
+        delete result;
+    }
+
+    // clear quest log tail
+    for (uint16 i = slot; i < MAX_QUEST_LOG_SIZE; ++i)
+    {
+        SetQuestSlot(i, 0);
+    }
+}
+
+void Player::_LoadDailyQuestStatus(QueryResult* result)
+{
+    for (uint32 quest_daily_idx = 0; quest_daily_idx < PLAYER_MAX_DAILY_QUESTS; ++quest_daily_idx)
+    {
+        SetUInt32Value(PLAYER_FIELD_DAILY_QUESTS_1 + quest_daily_idx, 0);
+    }
+
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `quest` FROM `character_queststatus_daily` WHERE `guid` = '%u'", GetGUIDLow());
+
+    if (result)
+    {
+        uint32 quest_daily_idx = 0;
+
+        do
+        {
+            if (quest_daily_idx >= PLAYER_MAX_DAILY_QUESTS) // max amount with exist data in query
+            {
+                sLog.outError("Player (GUID: %u) have more 25 daily quest records in `charcter_queststatus_daily`", GetGUIDLow());
+                break;
+            }
+
+            Field* fields = result->Fetch();
+
+            uint32 quest_id = fields[0].GetUInt32();
+
+            Quest const* pQuest = sObjectMgr.GetQuestTemplate(quest_id);
+            if (!pQuest)
+            {
+                continue;
+            }
+
+            SetUInt32Value(PLAYER_FIELD_DAILY_QUESTS_1 + quest_daily_idx, quest_id);
+            ++quest_daily_idx;
+
+            DEBUG_LOG("Daily quest {%u} cooldown for player (GUID: %u)", quest_id, GetGUIDLow());
+        }
+        while (result->NextRow());
+
+        delete result;
+    }
+
+    m_DailyQuestChanged = false;
+}
+
+void Player::_LoadWeeklyQuestStatus(QueryResult* result)
+{
+    m_weeklyquests.clear();
+
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT quest FROM character_queststatus_weekly WHERE guid = '%u'", GetGUIDLow());
+
+    if (result)
+    {
+        do
+        {
+            Field* fields = result->Fetch();
+
+            uint32 quest_id = fields[0].GetUInt32();
+
+            Quest const* pQuest = sObjectMgr.GetQuestTemplate(quest_id);
+            if (!pQuest)
+            {
+                continue;
+            }
+
+            m_weeklyquests.insert(quest_id);
+
+            DEBUG_LOG("Weekly quest {%u} cooldown for player (GUID: %u)", quest_id, GetGUIDLow());
+        }
+        while (result->NextRow());
+
+        delete result;
+    }
+    m_WeeklyQuestChanged = false;
+}
+
+void Player::_LoadMonthlyQuestStatus(QueryResult* result)
+{
+    m_monthlyquests.clear();
+
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT quest FROM character_queststatus_weekly WHERE guid = '%u'", GetGUIDLow());
+
+    if (result)
+    {
+        do
+        {
+            Field* fields = result->Fetch();
+
+            uint32 quest_id = fields[0].GetUInt32();
+
+            Quest const* pQuest = sObjectMgr.GetQuestTemplate(quest_id);
+            if (!pQuest)
+            {
+                continue;
+            }
+
+            m_monthlyquests.insert(quest_id);
+
+            DEBUG_LOG("Monthly quest {%u} cooldown for player (GUID: %u)", quest_id, GetGUIDLow());
+        }
+        while (result->NextRow());
+
+        delete result;
+    }
+
+    m_MonthlyQuestChanged = false;
+}
+
+/**
+ * @brief Loads known spells from the database.
+ *
+ * @param result The query result containing learned spell rows.
+ */
+void Player::_LoadSpells(QueryResult* result)
+{
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `spell`,`active`,`disabled` FROM `character_spell` WHERE `guid` = '%u'",GetGUIDLow());
+
+    if (result)
+    {
+        do
+        {
+            Field* fields = result->Fetch();
+
+            uint32 spell_id = fields[0].GetUInt32();
+
+            // skip talents & drop unneeded data
+            if (GetTalentSpellPos(spell_id))
+            {
+                sLog.outError("Player::_LoadSpells: %s has talent spell %u in character_spell, removing it.",
+                              GetGuidStr().c_str(), spell_id);
+                CharacterDatabase.PExecute("DELETE FROM `character_spell` WHERE `spell` = '%u'", spell_id);
+                continue;
+            }
+
+            addSpell(spell_id, fields[1].GetBool(), false, false, fields[2].GetBool());
+        }
+        while (result->NextRow());
+
+        delete result;
+    }
+}
+
+void Player::_LoadTalents(QueryResult* result)
+{
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT talent_id, current_rank, spec FROM character_talent WHERE guid = '%u'",GetGUIDLow());
+    if (result)
+    {
+        do
+        {
+            Field* fields = result->Fetch();
+
+            uint32 talent_id = fields[0].GetUInt32();
+            TalentEntry const* talentInfo = sTalentStore.LookupEntry(talent_id);
+
+            if (!talentInfo)
+            {
+                sLog.outError("Player::_LoadTalents:Player (GUID: %u) has invalid talent_id: %u , this talent will be deleted from character_talent", GetGUIDLow(), talent_id);
+                CharacterDatabase.PExecute("DELETE FROM `character_talent` WHERE `talent_id` = '%u'", talent_id);
+                continue;
+            }
+
+            TalentTabEntry const* talentTabInfo = sTalentTabStore.LookupEntry(talentInfo->TalentTab);
+
+            if (!talentTabInfo)
+            {
+                sLog.outError("Player::_LoadTalents:Player (GUID: %u) has invalid talentTabInfo: %u for talentID: %u , this talent will be deleted from character_talent", GetGUIDLow(), talentInfo->TalentTab, talentInfo->TalentID);
+                CharacterDatabase.PExecute("DELETE FROM `character_talent` WHERE `talent_id` = '%u'", talent_id);
+                continue;
+            }
+
+            // prevent load talent for different class (cheating)
+            if ((getClassMask() & talentTabInfo->ClassMask) == 0)
+            {
+                sLog.outError("Player::_LoadTalents:Player (GUID: %u) has talent with ClassMask: %u , but Player's ClassMask is: %u , talentID: %u , this talent will be deleted from character_talent", GetGUIDLow(), talentTabInfo->ClassMask, getClassMask() , talentInfo->TalentID);
+                CharacterDatabase.PExecute("DELETE FROM `character_talent` WHERE `guid` = '%u' AND `talent_id` = '%u'", GetGUIDLow(), talent_id);
+                continue;
+            }
+
+            uint32 currentRank = fields[1].GetUInt32();
+
+            if (currentRank > MAX_TALENT_RANK || talentInfo->RankID[currentRank] == 0)
+            {
+                sLog.outError("Player::_LoadTalents:Player (GUID: %u) has invalid talent rank: %u , talentID: %u , this talent will be deleted from character_talent", GetGUIDLow(), currentRank, talentInfo->TalentID);
+                CharacterDatabase.PExecute("DELETE FROM `character_talent` WHERE `guid` = '%u' AND `talent_id` = '%u'", GetGUIDLow(), talent_id);
+                continue;
+            }
+
+            uint32 spec = fields[2].GetUInt32();
+
+            if (spec > MAX_TALENT_SPEC_COUNT)
+            {
+                sLog.outError("Player::_LoadTalents:Player (GUID: %u) has invalid talent spec: %u, spec will be deleted from character_talent", GetGUIDLow(), spec);
+                CharacterDatabase.PExecute("DELETE FROM `character_talent` WHERE `spec` = '%u' ", spec);
+                continue;
+            }
+
+            if (spec >= m_specsCount)
+            {
+                sLog.outError("Player::_LoadTalents:Player (GUID: %u) has invalid talent spec: %u , this spec will be deleted from character_talent.", GetGUIDLow(), spec);
+                CharacterDatabase.PExecute("DELETE FROM `character_talent` WHERE `guid` = '%u' AND `spec` = '%u' ", GetGUIDLow(), spec);
+                continue;
+            }
+
+            if (m_activeSpec == spec)
+            {
+                addSpell(talentInfo->RankID[currentRank], true, false, false, false);
+            }
+            else
+            {
+                PlayerTalent talent;
+                talent.currentRank = currentRank;
+                talent.talentEntry = talentInfo;
+                talent.state       = PLAYERSPELL_UNCHANGED;
+                m_talents[spec][talentInfo->TalentID] = talent;
+            }
+        }
+        while (result->NextRow());
+        delete result;
+    }
+    UpdateGroupLeaderFlag();
+}
+
+/**
+ * @brief Loads the player's current group membership from the database.
+ *
+ * @param result The query result containing the group identifier.
+ */
+void Player::_LoadGroup(QueryResult* result)
+{
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `groupId` FROM `group_member` WHERE `memberGuid`='%u'", GetGUIDLow());
+    if (result)
+    {
+        uint32 groupId = (*result)[0].GetUInt32();
+        delete result;
+
+        if (Group* group = sObjectMgr.GetGroupById(groupId))
+        {
+            uint8 subgroup = group->GetMemberGroup(GetObjectGuid());
+            SetGroup(group, subgroup);
+            if (getLevel() >= LEVELREQUIREMENT_HEROIC)
+            {
+                // the group leader may change the instance difficulty while the player is offline
+                SetDungeonDifficulty(group->GetDungeonDifficulty());
+                SetRaidDifficulty(group->GetRaidDifficulty());
+            }
+        }
+    }
+}
+
+/**
+ * @brief Loads the player's saved dungeon and raid instance bindings.
+ *
+ * @param result The query result containing character instance bind rows.
+ */
+void Player::_LoadBoundInstances(QueryResult* result)
+{
+    for (uint8 i = 0; i < MAX_DIFFICULTY; ++i)
+    {
+        m_boundInstances[i].clear();
+    }
+
+    Group* group = GetGroup();
+
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `id`, `permanent`, `map`, `difficulty`, `resettime` FROM `character_instance` LEFT JOIN `instance` ON `instance` = `id` WHERE `guid` = '%u'", GUID_LOPART(m_guid));
+    if (result)
+    {
+        do
+        {
+            Field* fields = result->Fetch();
+            bool perm = fields[1].GetBool();
+            uint32 mapId = fields[2].GetUInt32();
+            uint32 instanceId = fields[0].GetUInt32();
+            uint8 difficulty = fields[3].GetUInt8();
+
+            time_t resetTime = (time_t)fields[4].GetUInt64();
+            // the resettime for normal instances is only saved when the InstanceSave is unloaded
+            // so the value read from the DB may be wrong here but only if the InstanceSave is loaded
+            // and in that case it is not used
+
+            MapEntry const* mapEntry = sMapStore.LookupEntry(mapId);
+            if (!mapEntry || !mapEntry->IsDungeon())
+            {
+                sLog.outError("_LoadBoundInstances: player %s(%d) has bind to nonexistent or not dungeon map %d", GetName(), GetGUIDLow(), mapId);
+                CharacterDatabase.PExecute("DELETE FROM `character_instance` WHERE `guid` = '%u' AND `instance` = '%u'", GetGUIDLow(), instanceId);
+                continue;
+            }
+
+            if (difficulty >= MAX_DIFFICULTY)
+            {
+                sLog.outError("_LoadBoundInstances: player %s(%d) has bind to nonexistent difficulty %d instance for map %u", GetName(), GetGUIDLow(), difficulty, mapId);
+                CharacterDatabase.PExecute("DELETE FROM `character_instance` WHERE `guid` = '%u' AND `instance` = '%u'", GetGUIDLow(), instanceId);
+                continue;
+            }
+
+            MapDifficultyEntry const* mapDiff = GetMapDifficultyData(mapId, Difficulty(difficulty));
+            if (!mapDiff)
+            {
+                sLog.outError("_LoadBoundInstances: player %s(%d) has bind to nonexistent difficulty %d instance for map %u", GetName(), GetGUIDLow(), difficulty, mapId);
+                CharacterDatabase.PExecute("DELETE FROM `character_instance` WHERE `guid` = '%u' AND `instance` = '%u'", GetGUIDLow(), instanceId);
+                continue;
+            }
+
+            if (!perm && group)
+            {
+                sLog.outError("_LoadBoundInstances: %s is in group (Id: %d) but has a non-permanent character bind to map %d,%d,%d",
+                              GetGuidStr().c_str(), group->GetId(), mapId, instanceId, difficulty);
+                CharacterDatabase.PExecute("DELETE FROM `character_instance` WHERE `guid` = '%u' AND `instance` = '%u'",
+                                           GetGUIDLow(), instanceId);
+                continue;
+            }
+
+            // since non permanent binds are always solo bind, they can always be reset
+            DungeonPersistentState* state = reinterpret_cast<DungeonPersistentState*>(sMapPersistentStateMgr.AddPersistentState(mapEntry, instanceId, Difficulty(difficulty), resetTime, !perm, true));
+            if (state)
+            {
+                BindToInstance(state, perm, true);
+            }
+        }
+        while (result->NextRow());
+        delete result;
+    }
+}
+
+/**
+ * @brief Finds the player's instance bind for a specific map.
+ *
+ * @param mapid The map identifier to look up.
+ * @return The matching bind entry, or NULL if none exists.
+ */
+InstancePlayerBind* Player::GetBoundInstance(uint32 mapid, Difficulty difficulty)
+{
+    // some instances only have one difficulty
+    MapDifficultyEntry const* mapDiff = GetMapDifficultyData(mapid, difficulty);
+    if (!mapDiff)
+    {
+        return NULL;
+    }
+
+    BoundInstancesMap::iterator itr = m_boundInstances[difficulty].find(mapid);
+    if (itr != m_boundInstances[difficulty].end())
+    {
+        return &itr->second;
+    }
+    else
+    {
+        return NULL;
+    }
+}
+
+/**
+ * @brief Removes the player's instance bind for a map.
+ *
+ * @param mapid The bound map identifier.
+ * @param unload True when called during load or unload paths that should skip DB deletion.
+ */
+void Player::UnbindInstance(uint32 mapid, Difficulty difficulty, bool unload)
+{
+    BoundInstancesMap::iterator itr = m_boundInstances[difficulty].find(mapid);
+    UnbindInstance(itr, difficulty, unload);
+}
+
+/**
+ * @brief Removes the instance bind referenced by an iterator.
+ *
+ * @param itr Iterator pointing at the bind to remove.
+ * @param unload True when called during load or unload paths that should skip DB deletion.
+ */
+void Player::UnbindInstance(BoundInstancesMap::iterator& itr, Difficulty difficulty, bool unload)
+{
+    if (itr != m_boundInstances[difficulty].end())
+    {
+        if (!unload)
+            CharacterDatabase.PExecute("DELETE FROM `character_instance` WHERE `guid` = '%u' AND `instance` = '%u'",
+                                       GetGUIDLow(), itr->second.state->GetInstanceId());
+
+        sCalendarMgr.SendCalendarRaidLockoutRemove(this, itr->second.state);
+
+        itr->second.state->RemovePlayer(this);              // state can become invalid
+        m_boundInstances[difficulty].erase(itr++);
+    }
+}
+
+/**
+ * @brief Binds the player to a dungeon persistent state.
+ *
+ * @param state The persistent instance state to bind.
+ * @param permanent True for a permanent lock, false for a temporary one.
+ * @param load True when rebuilding the bind from saved data.
+ * @return The resulting bind entry, or NULL if the state was invalid.
+ */
+InstancePlayerBind* Player::BindToInstance(DungeonPersistentState* state, bool permanent, bool load)
+{
+    if (state)
+    {
+        InstancePlayerBind& bind = m_boundInstances[state->GetDifficulty()][state->GetMapId()];
+        if (bind.state)
+        {
+            // update the state when the group kills a boss
+            if (permanent != bind.perm || state != bind.state)
+                if (!load)
+                    CharacterDatabase.PExecute("UPDATE `character_instance` SET `instance` = '%u', `permanent` = '%u' WHERE `guid` = '%u' AND `instance` = '%u'",
+                                               state->GetInstanceId(), permanent, GetGUIDLow(), bind.state->GetInstanceId());
+        }
+        else
+        {
+            if (!load)
+                CharacterDatabase.PExecute("INSERT INTO `character_instance` (`guid`, `instance`, `permanent`) VALUES ('%u', '%u', '%u')",
+                                           GetGUIDLow(), state->GetInstanceId(), permanent);
+        }
+
+        if (bind.state != state)
+        {
+            if (bind.state)
+            {
+                bind.state->RemovePlayer(this);
+            }
+            state->AddPlayer(this);
+        }
+
+        if (permanent)
+        {
+            state->SetCanReset(false);
+        }
+
+        bind.state = state;
+        bind.perm = permanent;
+        if (!load)
+        {
+            DEBUG_LOG("Player::BindToInstance: %s(%d) is now bound to map %d, instance %d, difficulty %d",
+                      GetName(), GetGUIDLow(), state->GetMapId(), state->GetInstanceId(), state->GetDifficulty());
+        }
+        // Used by Eluna
+#ifdef ENABLE_ELUNA
+        if (Eluna* e = GetEluna())
+        {
+            e->OnBindToInstance(this, (Difficulty)0, state->GetMapId(), permanent);
+        }
+#endif /* ENABLE_ELUNA */
+        return &bind;
+    }
+    else
+    {
+        return NULL;
+    }
+}
+
+/**
+ * @brief Resolves the instance save that applies to the player or the player's group.
+ *
+ * @param mapid The map identifier to inspect.
+ * @return The applicable persistent state, or NULL if none exists.
+ */
+DungeonPersistentState* Player::GetBoundInstanceSaveForSelfOrGroup(uint32 mapid)
+{
+    MapEntry const* mapEntry = sMapStore.LookupEntry(mapid);
+    if (!mapEntry)
+    {
+        return NULL;
+    }
+
+    InstancePlayerBind* pBind = GetBoundInstance(mapid, GetDifficulty(mapEntry->IsRaid()));
+    DungeonPersistentState* state = pBind ? pBind->state : NULL;
+
+    // the player's permanent player bind is taken into consideration first
+    // then the player's group bind and finally the solo bind.
+    if (!pBind || !pBind->perm)
+    {
+        InstanceGroupBind* groupBind = NULL;
+        // use the player's difficulty setting (it may not be the same as the group's)
+        if (Group* group = GetGroup())
+            if ((groupBind = group->GetBoundInstance(mapid, this)))
+            {
+                state = groupBind->state;
+            }
+    }
+
+    return state;
+}
+
+/**
+ * @brief Sends the player's permanent raid instance lock information to the client.
+ */
+void Player::SendRaidInfo()
+{
+    uint32 counter = 0;
+
+    WorldPacket data(SMSG_RAID_INSTANCE_INFO, 4);
+
+    size_t p_counter = data.wpos();
+    data << uint32(counter);                                // placeholder
+
+    time_t now = time(NULL);
+
+    for (int i = 0; i < MAX_DIFFICULTY; ++i)
+    {
+        for (BoundInstancesMap::const_iterator itr = m_boundInstances[i].begin(); itr != m_boundInstances[i].end(); ++itr)
+        {
+            if (itr->second.perm)
+            {
+                DungeonPersistentState* state = itr->second.state;
+                data << uint32(state->GetMapId());          // map id
+                data << uint32(state->GetDifficulty());     // difficulty
+                data << ObjectGuid(state->GetInstanceGuid());// instance guid
+                data << uint8(1);                           // expired = 0
+                data << uint8(0);                           // extended = 1
+                data << uint32(state->GetResetTime() - now);// reset time
+                data << uint32(state->GetCompletedEncountersMask());// completed encounter mask
+                ++counter;
+            }
+        }
+    }
+    data.put<uint32>(p_counter, counter);
+    GetSession()->SendPacket(&data);
+}
+
+/*
+- called on every successful teleportation to a map
+*/
+void Player::SendSavedInstances()
+{
+    bool hasBeenSaved = false;
+    WorldPacket data;
+
+    for (uint8 i = 0; i < MAX_DIFFICULTY; ++i)
+    {
+        for (BoundInstancesMap::const_iterator itr = m_boundInstances[i].begin(); itr != m_boundInstances[i].end(); ++itr)
+        {
+            if (itr->second.perm)                           // only permanent binds are sent
+            {
+                hasBeenSaved = true;
+                break;
+            }
+        }
+    }
+
+    // Send opcode 811. true or false means, whether you have current raid/heroic instances
+    data.Initialize(SMSG_UPDATE_INSTANCE_OWNERSHIP);
+    data << uint32(hasBeenSaved);
+    GetSession()->SendPacket(&data);
+
+    if (!hasBeenSaved)
+    {
+        return;
+    }
+
+    for (uint8 i = 0; i < MAX_DIFFICULTY; ++i)
+    {
+        for (BoundInstancesMap::const_iterator itr = m_boundInstances[i].begin(); itr != m_boundInstances[i].end(); ++itr)
+        {
+            if (itr->second.perm)
+            {
+                data.Initialize(SMSG_UPDATE_LAST_INSTANCE);
+                data << uint32(itr->second.state->GetMapId());
+                GetSession()->SendPacket(&data);
+            }
+        }
+    }
+}
+
+/// convert the player's binds to the group
+void Player::ConvertInstancesToGroup(Player* player, Group* group, ObjectGuid player_guid)
+{
+    bool has_binds = false;
+    bool has_solo = false;
+
+    if (player)
+    {
+        player_guid = player->GetObjectGuid();
+        if (!group)
+        {
+            group = player->GetGroup();
+        }
+    }
+
+    MANGOS_ASSERT(player_guid);
+
+    // copy all binds to the group, when changing leader it's assumed the character
+    // will not have any solo binds
+
+    if (player)
+    {
+        for (uint8 i = 0; i < MAX_DIFFICULTY; ++i)
+        {
+            for (BoundInstancesMap::iterator itr = player->m_boundInstances[i].begin(); itr != player->m_boundInstances[i].end();)
+            {
+                has_binds = true;
+
+                if (group)
+                {
+                    group->BindToInstance(itr->second.state, itr->second.perm, true);
+                }
+
+                // permanent binds are not removed
+                if (!itr->second.perm)
+                {
+                    // increments itr in call
+                    player->UnbindInstance(itr, Difficulty(i), true);
+                    has_solo = true;
+                }
+                else
+                {
+                    ++itr;
+                }
+            }
+        }
+    }
+
+    uint32 player_lowguid = player_guid.GetCounter();
+
+    // if the player's not online we don't know what binds it has
+    if (!player || !group || has_binds)
+    {
+        CharacterDatabase.PExecute("INSERT INTO `group_instance` SELECT `guid`, `instance`, `permanent` FROM `character_instance` WHERE `guid` = '%u'", player_lowguid);
+    }
+
+    // the following should not get executed when changing leaders
+    if (!player || has_solo)
+    {
+        CharacterDatabase.PExecute("DELETE FROM `character_instance` WHERE `guid` = '%u' AND `permanent` = 0", player_lowguid);
+    }
+}
+
+/**
+ * @brief Loads and validates the player's home bind location.
+ *
+ * @param result The query result containing home bind data.
+ * @return True if a valid home bind was loaded or defaulted successfully; otherwise, false.
+ */
+bool Player::_LoadHomeBind(QueryResult* result)
+{
+    PlayerInfo const* info = sObjectMgr.GetPlayerInfo(getRace(), getClass());
+    if (!info)
+    {
+        sLog.outError("Player have incorrect race/class pair. Can't be loaded.");
+        return false;
+    }
+
+    bool ok = false;
+    // QueryResult *result = CharacterDatabase.PQuery("SELECT `map`,`zone`,`position_x`,`position_y`,`position_z` FROM `character_homebind` WHERE `guid` = '%u'", GUID_LOPART(playerGuid));
+    if (result)
+    {
+        Field* fields = result->Fetch();
+        m_homebindMapId = fields[0].GetUInt32();
+        m_homebindAreaId = fields[1].GetUInt16();
+        m_homebindX = fields[2].GetFloat();
+        m_homebindY = fields[3].GetFloat();
+        m_homebindZ = fields[4].GetFloat();
+        delete result;
+
+        MapEntry const* bindMapEntry = sMapStore.LookupEntry(m_homebindMapId);
+
+        // accept saved data only for valid position (and non instanceable), and accessable
+        if (MapManager::IsValidMapCoord(m_homebindMapId, m_homebindX, m_homebindY, m_homebindZ) &&
+                !bindMapEntry->Instanceable() && GetSession()->Expansion() >= bindMapEntry->Expansion())
+        {
+            ok = true;
+        }
+        else
+        {
+            CharacterDatabase.PExecute("DELETE FROM `character_homebind` WHERE `guid` = '%u'", GetGUIDLow());
+        }
+    }
+
+    if (!ok)
+    {
+        m_homebindMapId = info->mapId;
+        m_homebindAreaId = info->areaId;
+        m_homebindX = info->positionX;
+        m_homebindY = info->positionY;
+        m_homebindZ = info->positionZ;
+
+        CharacterDatabase.PExecute("INSERT INTO `character_homebind` (`guid`,`map`,`zone`,`position_x`,`position_y`,`position_z`) VALUES ('%u', '%u', '%u', '%f', '%f', '%f')", GetGUIDLow(), m_homebindMapId, (uint32)m_homebindAreaId, m_homebindX, m_homebindY, m_homebindZ);
+    }
+
+    DEBUG_LOG("Setting player home position: mapid is: %u, zoneid is %u, X is %f, Y is %f, Z is %f",
+              m_homebindMapId, m_homebindAreaId, m_homebindX, m_homebindY, m_homebindZ);
+
+    return true;
+}
+
